requirements.txt
anyio==3.6.2
autopep8==2.0.0
click==8.1.3
defusedxml==0.7.1
fastapi==0.87.0
fonttools==4.38.0
fpdf2==2.6.0
h11==0.14.0
idna==3.4
Pillow==9.3.0
pycodestyle==2.9.1
pydantic==1.10.2
sniffio==1.3.0
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
